:hammer: refactor: app 이름 변경 및 변수명 변경
앱 이름: app -> chat_app
변수명: Consumer::room -> chatroom
