+++ chat_app/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'app'
+    name = 'chat_app'
