+++ config/asgi.py
@@ -6,7 +6,7 @@ from channels.routing import ProtocolTypeRouter, URLRouter
 django_asgi_app = get_asgi_application()
 
 from channels.auth import AuthMiddlewareStack
-from app.routing import websocket_urlpatterns
+from chat_app.routing import websocket_urlpatterns
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')
