+++ config/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path, include
 
 urlpatterns = [
-    path('api/chat/', include('app.urls')),
+    path('api/chat/', include('chat_app.urls')),
 ]
